Re-factored original code
