+++ input/6C2H4.txt
@@ -1,0 +1,37 @@
+36 0
+6       1.266     0.000      0.0
+6      -1.266    -0.000      0.0
+1       2.305     1.800      0.0
+1       2.305    -1.800      0.0
+1      -2.305     1.800      0.0
+1      -2.305    -1.800      0.0
+6       1.266     0.000      3.0
+6      -1.266    -0.000      3.0
+1       2.305     1.800      3.0
+1       2.305    -1.800      3.0
+1      -2.305     1.800      3.0
+1      -2.305    -1.800      3.0
+6       1.266     0.000      6.0
+6      -1.266    -0.000      6.0
+1       2.305     1.800      6.0
+1       2.305    -1.800      6.0
+1      -2.305     1.800      6.0
+1      -2.305    -1.800      6.0
+6       1.266     0.000      9.0
+6      -1.266    -0.000      9.0
+1       2.305     1.800      9.0
+1       2.305    -1.800      9.0
+1      -2.305     1.800      9.0
+1      -2.305    -1.800      9.0
+6       1.266     0.000      12.0
+6      -1.266    -0.000      12.0
+1       2.305     1.800      12.0
+1       2.305    -1.800      12.0
+1      -2.305     1.800      12.0
+1      -2.305    -1.800      12.0
+6       1.266     0.000      15.0
+6      -1.266    -0.000      15.0
+1       2.305     1.800      15.0
+1       2.305    -1.800      15.0
+1      -2.305     1.800      15.0
+1      -2.305    -1.800      15.0
