+++ input/C10H22.txt
@@ -1,10 +1,14 @@
-20 0
+32 0
 6 0.0 0.0 0.0
 6 1.2576 0.8888 0.0
 6 2.5152 0.0 0.0
 6 3.7728 0.8888 0.0
 6 5.0304 0.0 0.0
 6 6.288 0.8888 0.0
+6 7.5456 0.0 0.0
+6 8.8032 0.8888 0.0
+6 10.0608 0.0 0.0
+1 11.3184 0.8888 0.0
 1 -0.89 0.6291 0.0
 1 0.0 -0.6291 0.89
 1 0.0 -0.6291 -0.89
@@ -16,6 +20,14 @@
 1 3.7728 1.5179 -0.89
 1 5.0304 -0.6291 0.89
 1 5.0304 -0.6291 -0.89
-1 7.178 1.5179 0.0
-1 6.288 0.25970000000000004 0.89
-1 6.288 0.25970000000000004 -0.89
+1 6.288 1.5179 0.89
+1 6.288 1.5179 -0.89
+1 7.5456 -0.6291 0.89
+1 7.5456 -0.6291 -0.89
+1 8.8032 1.5179 0.89
+1 8.8032 1.5179 -0.89
+1 10.0608 -0.6291 0.89
+1 10.0608 -0.6291 -0.89
+1 12.2084 1.5179 0.0
+1 11.3184 0.2597 0.89
+1 11.3184 0.2597 -0.89
