+++ input/C200H402.txt
@@ -1,4 +1,4 @@
-602 0 601 601
+602 0
 6 0.0 0.0 0.0
 6 1.2576 0.8888 0.0
 6 2.5152 0.0 0.0
