+++ input/C5H12.txt
@@ -1,4 +1,4 @@
-17 0 16 16
+17 0
 6 0.0 0.0 0.0
 6 1.2576 0.8888 0.0
 6 2.5152 0.0 0.0
