+++ input/C7H16.txt
@@ -1,4 +1,4 @@
-23 0 22 22
+23 0
 6 0.0 0.0 0.0
 6 1.2576 0.8888 0.0
 6 2.5152 0.0 0.0
