+++ input/C7H16_copy.txt
@@ -1,0 +1,24 @@
+23 0
+C 0.0 0.0 0.0
+C 1.2576 0.8888 0.0
+C 2.5152 0.0 0.0
+C 3.7728 0.8888 0.0
+C 5.0304 0.0 0.0
+C 6.288 0.8888 0.0
+C 7.5456 0.0 0.0
+H -0.89 0.6291 0.0
+H 0.0 -0.6291 0.89
+H 0.0 -0.6291 -0.89
+H 1.2576 1.5179 0.89
+H 1.2576 1.5179 -0.89
+H 2.5152 -0.6291 0.89
+H 2.5152 -0.6291 -0.89
+H 3.7728 1.5179 0.89
+H 3.7728 1.5179 -0.89
+H 5.0304 -0.6291 0.89
+H 5.0304 -0.6291 -0.89
+H 6.288 1.5179 0.89
+H 6.288 1.5179 -0.89
+H 8.4356 0.6291 0.0
+H 7.5456 -0.6291 0.89
+H 7.5456 -0.6291 -0.89
