+++ src/main.cpp
@@ -20,8 +20,22 @@ int main(int argc, char* argv[])
   // Parse input file, create molecule
   try {
     Molecule my_molecule = read_molecule(fname);
+    // // To print information about each basis function
+    // vector<Atom> atoms = my_molecule.get_atoms();
+    // for (int i = 0; i < atoms.size(); i++) {
+    //   vector<BasisFunction>& basis_functions = atoms[i].get_basis_functions();
+    //   for (int j = 0; j < basis_functions.size(); j++) {
+    //     basis_functions[j].print_info();
+    //   }
+    // }
 
-    // To print information about each basis function
+    // for (int i = 0; i < my_molecule.get_all_basis_functions().size(); i++) {
+    //    my_molecule.get_all_basis_functions()[i]->print_info();
+    // }
+
+    //my_molecule.write_output("output.txt");
+
+    // // To print information about each basis function
     // vector<Atom> atoms = my_molecule.get_atoms();
     // for (int i = 0; i < atoms.size(); i++) {
     //   vector<BasisFunction>& basis_functions = atoms[i].get_basis_functions();
