--- src/test.txt
@@ -1,108 +1,0 @@
-      0   0.3571  -0.1807        0        0  
--0.3571        0        0        0        0   
- 0.1807        0        0        0        0
-       0        0        0        0        0        
-       0        0        0        0        0
-
-0        0        0   0.2044        0        
-0        0        0        0        0        
-0        0        0        0        0  
--0.2044        0        0        0        0        
-0        0        0        0        0
-
-0 1 0 1 0.00183174
-0 1 0 1 0.0832633
-0 1 0 1 0.298559
-0 1 0 1 0.107271
-0 1 0 1 0.739203
-0 1 0 1 1.98859
-0 1 0 1 0.113382
-0 1 0 1 0.858998
-0 1 0 1 3.11076
-0 2 0 1 -0.00100879
-0 2 0 1 -0.083439
-0 2 0 1 -0.27178
-0 2 0 1 -0.0114158
-0 2 0 1 -0.230051
-0 2 0 1 -0.735056
-0 2 0 1 0.000134329
-0 2 0 1 0.0169321
-0 2 0 1 0.476619
-
-1 0 1 0 0.00183174
-1 0 1 0 0.0832633
-1 0 1 0 0.298559
-1 0 1 0 0.107271
-1 0 1 0 0.739203
-1 0 1 0 1.98859
-1 0 1 0 0.113382
-1 0 1 0 0.858998
-1 0 1 0 3.11076
-2 0 1 0 -0.00100879
-2 0 1 0 -0.083439
-2 0 1 0 -0.27178
-2 0 1 0 -0.0114158
-2 0 1 0 -0.230051
-2 0 1 0 -0.735056
-2 0 1 0 0.000134329
-2 0 1 0 0.0169321
-2 0 1 0 0.476619
-
-0 1
-0 1 -0.000146529
-0 1 0.00890911
-0 1 0.0241098
-0 1 -0.00829929
-0 1 0.0764965
-0 1 0.155313
-0 1 -0.00273387
-0 1 0.0277042
-0 1 0.0757187
-0.357071
-0 2
-0 1 -0.000640032
-0 1 -0.0332892
-0 1 -0.0171772
-0 1 -0.00700494
-0 1 -0.0887676
-0 1 -0.0449316
-0 1 2.56888e-05
-0 1 0.00203619
-0 1 0.00907985
--0.180669
-
-1 0
-1 0 -0.000146529
-1 0 -0.00644186
-1 0 -0.00719886
-1 0 0.011478
-1 0 0.0764965
-1 0 0.0641357
-1 0 0.00915606
-1 0 0.0670893
-1 0 0.0757187
--0.290287
-2 0
-1 0 -0.000640032
-1 0 -0.0511998
-1 0 -0.0519749
-1 0 -0.00455448
-1 0 -0.0887676
-1 0 -0.0883948
-1 0 8.48991e-06
-1 0 0.00103501
-1 0 0.00907985
-0.275408
-
-0            0            0            0  -2.7696e-01  -1.8561e-01            0            0   1.5144e-01   1.5144e-01  -8.9084e-02  -8.9084e-02            
-0            0            0            0   1.8561e-01  -1.6771e-02            0            0  -1.2145e-01  -1.2145e-01   8.1485e-02   8.1485e-02            
-0            0            0            0            0            0  -2.2333e-01            0   1.7627e-01  -1.7627e-01  -5.7913e-02   5.7913e-02            
-0            0            0            0            0            0            0  -2.2333e-01            0            0            0            0   
-2.7696e-01  -1.8561e-01            0            0            0            0            0            0   8.9084e-02   8.9084e-02  -1.5144e-01  -1.5144e-01   
-1.8561e-01   1.6771e-02            0            0            0            0            0            0   8.1485e-02   8.1485e-02  -1.2145e-01  -1.2145e-01            
-0            0   2.2333e-01            0            0            0            0            0   5.7913e-02  -5.7913e-02  -1.7627e-01   1.7627e-01            
-0            0            0   2.2333e-01            0            0            0            0            0            0            0            0  
--1.5144e-01   1.2145e-01  -1.7627e-01            0  -8.9084e-02  -8.1485e-02  -5.7913e-02            0            0            0  -5.4167e-02  -1.3396e-02  
--1.5144e-01   1.2145e-01   1.7627e-01            0  -8.9084e-02  -8.1485e-02   5.7913e-02            0            0            0  -1.3396e-02  -5.4167e-02   
-8.9084e-02  -8.1485e-02   5.7913e-02            0   1.5144e-01   1.2145e-01   1.7627e-01            0   5.4167e-02   1.3396e-02            0            0   
-8.9084e-02  -8.1485e-02  -5.7913e-02            0   1.5144e-01   1.2145e-01  -1.7627e-01            0   1.3396e-02   5.4167e-02            0            0
